refactor: move common/ into shared/
