--- src/common/types.ts
@@ -1,38 +1,0 @@
-export type Account = {
-  password: string;
-  username: string;
-};
-
-export type AccountWithServer = Account & {
-  server: string | null;
-};
-
-export type FastTravel = {
-  /**
-   * The cell to jump to. Defaults to "Enter".
-   */
-  cell?: string;
-  /**
-   * The map name to join.
-   */
-  map: string;
-  /**
-   * The display name of the location.
-   */
-  name: string;
-  /**
-   * The pad to jump to. Defaults to "Spawn".
-   */
-  pad?: string;
-};
-
-export type Settings = {
-  /**
-   * Whether debug logging is enabled.
-   */
-  debug: boolean;
-  /**
-   * The launch mode of the application.
-   */
-  launchMode: "game" | "manager";
-};
