+++ src/main/index.ts
@@ -4,9 +4,9 @@ import { join } from "path";
 import process from "process";
 import { registerIpcMain } from "@egoist/tipc/main";
 import { app } from "electron";
-import { FileManager } from "../common/FileManager";
-import { BRAND } from "../common/constants";
-import type { Settings } from "../common/types";
+import { FileManager } from "../shared/FileManager";
+import { BRAND } from "../shared/constants";
+import type { Settings } from "../shared/types";
 import { router } from "./tipc";
 import { showErrorDialog } from "./util/showErrorDialog";
 import { createAccountManager, createGame } from "./windows";
