+++ src/renderer/game/lib/Settings.ts
@@ -1,4 +1,4 @@
-import { interval } from "../../../common/interval";
+import { interval } from "../../../shared/interval";
 import { gameState } from "../state.svelte";
 import type { Bot } from "./Bot";
