+++ src/shared/FileManager.ts
