+++ src/shared/index.ts
@@ -1,3 +1,7 @@
 export * from "./sleep";
 export * from "./cn";
 export type * from "./types";
+export * from "./FileManager";
+export * from "./constants";
+export * from "./tipc";
+export * from "./logger";
