+++ src/shared/interval.ts
